Move cast_parameter_value over to tests
